Feat (axe): improved implementation with extended support
