+++ src/brevitas/graph/gpxq.py
@@ -7,17 +7,23 @@ from copy import deepcopy
 from dataclasses import dataclass
 from dataclasses import field
 from functools import partial
+import math
 from operator import attrgetter
 from typing import List
 from typing import Optional
 from typing import Set
+from typing import Union
 import warnings
 
+import numpy as np
 import torch
+from torch import Tensor
 from torch.fx import GraphModule as TorchGraphModule
 import torch.nn as nn
 import unfoldNd
 
+from brevitas.function.ops import max_int
+from brevitas.function.ops import min_int
 from brevitas.fx import GraphModule
 from brevitas.graph.calibrate import quantization_status_manager
 from brevitas.graph.utils import is_conv_transposed
@@ -25,6 +31,7 @@ from brevitas.graph.utils import is_quant_module
 import brevitas.nn as qnn
 from brevitas.quant_tensor import _unpack_quant_tensor
 from brevitas.quant_tensor import QuantTensor
+from brevitas.utils.quant_utils import _CachedIO, _CachedIOGroupwiseInt
 
 SUPPORTED_CONV_OP = (
     nn.Conv1d, nn.Conv2d, nn.Conv3d, nn.ConvTranspose1d, nn.ConvTranspose2d, nn.ConvTranspose3d)
@@ -386,3 +393,154 @@ class GPxQ(ABC):
         # We need to remove the last dim
         q = q.squeeze(2)  # [groups, OC/groups] or [1, OC]
         return q
+
+
+def _get_average_of_nonzero_magnitudes(vec: np.ndarray, radius: float = 1.0):
+    assert radius > 0, "Error: radius needs to be strictly positive."
+    assert vec.ndim == 1, "Error: projection assumes a vector, not a matrix."
+    assert vec.min() >= 0, "Error: assuming a vector of non-negative numbers."
+    n_elems = vec.shape[0]
+    # if we are already within the simplex, then the best projection is itself
+    if vec.sum() <= radius:
+        return 0.0
+    # using algorithm detailed in "Efficient Projections onto the L1-Ball for Learning in High Dimensions"
+    v = vec
+    u = np.sort(v)[::-1]
+    cumsum_u = np.cumsum(u)
+    rho = np.nonzero(u * np.arange(1, n_elems + 1) > (cumsum_u - radius))[0][-1]
+    theta = float(cumsum_u[rho] - radius) / (rho + 1)
+    return theta
+
+
+def calc_average_nonzero_mag(weight: Tensor, lim: Tensor) -> Tensor:
+    thetas = torch.zeros(weight.shape[0], device=weight.device)
+    for i in range(weight.shape[0]):
+        l = lim[i].item() if lim.ndim > 0 else lim.item()
+        w = weight[i].cpu().detach().numpy()
+        t = _get_average_of_nonzero_magnitudes(np.abs(w), l)
+        thetas[i] = t
+    return thetas
+
+
+def pad_tensor_with_zeros(tensor: Tensor, tile_size: int) -> Tensor:
+    pad_size = tile_size - (tensor.shape[1] % tile_size)
+    if pad_size == tile_size:
+        return tensor
+    padding = torch.zeros((tensor.shape[0], pad_size), device=tensor.device)
+    pad_tensor = torch.concat([tensor, padding], axis=1)
+    return pad_tensor
+
+
+class AXE:
+    """
+    Accumulator-aware extensions for greedy path sequential quantization algorithms
+    such as the GPxQ family of algorithms.
+
+    See "Accumulator-Aware Post-Training Quantization" for more details.
+    """
+
+    quant_metadata: Union[_CachedIO, _CachedIOGroupwiseInt] = None
+    max_accumulator_bit_width: Tensor = None
+    max_accumulator_tile_size: int = None
+
+    @property
+    def input_min(self):
+        assert self.quant_metadata is not None, "Error: need quantized activations"
+        input_bit_width = self.quant_metadata.bit_width
+        input_is_signed = self.quant_metadata.signed
+        # NOTE: can't get this from cache, so assuming worst-case scenario
+        input_is_narrow = False
+        input_min = min_int(input_is_signed, input_is_narrow, input_bit_width)
+        assert input_min <= 0, f"Error: input_min={input_min}. Should be non-positive."
+        return int(input_min)
+
+    @property
+    def input_max(self):
+        assert self.quant_metadata is not None, "Error: need quantized activations"
+        input_bit_width = self.quant_metadata.bit_width
+        input_is_signed = self.quant_metadata.signed
+        # NOTE: can't get this from cache, so assuming worst-case scenario
+        input_is_narrow = False
+        input_max = max_int(input_is_signed, input_is_narrow, input_bit_width)
+        assert input_max >= 0, f"Error: input_max={input_max}. Should be non-negative."
+        return int(input_max)
+
+    def upper_lim(self, n: Tensor, p: Tensor):
+        p0 = torch.exp2(self.max_accumulator_bit_width - 1.) - 1.
+        p1 = (self.input_max * p) + (self.input_min * n)
+        p2 = (p0 - p1) / self.input_max
+        assert (p2 >= 0).all()
+
+        # for unsigned data types, assuming round-to-nearest
+        if self.input_min == 0:
+            return p2 - 0.5
+
+        n0 = -torch.exp2(self.max_accumulator_bit_width - 1.) + 1.
+        n1 = (self.input_min * p) + (self.input_max * n)
+        n2 = (n0 - n1) / self.input_min
+        assert (n2 >= 0).all()
+
+        # take the most restrictive lower limit (i.e., the smallest one),
+        # note that we are assuming round-to-nearest here
+        return torch.where(p2 < n2, p2, n2) - 0.5
+
+    def lower_lim(self, n: Tensor, p: Tensor):
+        n0 = -torch.exp2(self.max_accumulator_bit_width - 1.) + 1.
+        n1 = (self.input_min * p) + (self.input_max * n)
+        n2 = (n0 - n1) / self.input_max
+        assert (n2 <= 0).all()
+
+        # for unsigned data types, assuming round-to-nearest
+        if self.input_min == 0:
+            return n2 + 0.5
+
+        p0 = torch.exp2(self.max_accumulator_bit_width - 1.) - 1.
+        p1 = (self.input_max * p) + (self.input_min * n)
+        p2 = (p0 - p1) / self.input_min
+        assert (p2 <= 0).all()
+
+        # take the most restrictive lower limit (i.e., the largest one),
+        # note that we are assuming round-to-nearest here
+        return torch.where(p2 > n2, p2, n2) + 0.5
+
+    def get_scales_and_thresholds(self, weight: Tensor):
+        # NOTE: assuming sign-magnitude here, which is sufficient to support both
+        # sign-magnitude and 2s complement accumulators
+        Z = (torch.exp2(self.max_accumulator_bit_width) -
+             2) / float(self.input_max - self.input_min)
+        n_tiles = math.ceil(weight.shape[-1] / self.max_accumulator_tile_size)
+
+        scales = self.layer.weight_quant.scale()
+        if scales.ndim > 0:
+            if isinstance(self.layer, SUPPORTED_CONV_OP):
+                if isinstance(self.layer, SUPPORTED_TCONV_OP):
+                    scales = scales.transpose(1, 0)  # This performs a view
+                scales = scales.flatten(1)
+
+        # translating into the quantized range; need to pad to get these thresholds
+        wT = pad_tensor_with_zeros(weight / scales, self.max_accumulator_tile_size).view(
+            -1, self.max_accumulator_tile_size)  # [OC * Tiles, IC / Tiles]
+        # calculate the thresholds after zero centering projection
+        thresholds = calc_average_nonzero_mag(
+            wT - wT.mean(axis=1, keepdim=True), Z)  # [Groups * OC * Tiles]
+        thresholds = thresholds.view(self.groups, -1,
+                                     n_tiles).transpose(1, 2)  # [Groups, Tiles, OC/Groups]
+        del wT
+        # supporting groupwise quantization where each tile has its own scaling factor
+        if self.layer.weight_quant.is_groupwise:
+            if (self.max_accumulator_tile_size != self.columns) and (self.max_accumulator_tile_size != self.layer.weight_quant.group_size):
+                raise ValueError(
+                    "Error: only supporting accumulator-aware groupwise weight quantization"
+                    "when the group size is equal to the accumulator tile size or a monolithic" 
+                    "accumulator is assumed (i.e., `max_accumulator_tile_size=None`).")
+            scales = pad_tensor_with_zeros(scales, self.max_accumulator_tile_size).view(
+                -1, self.max_accumulator_tile_size)  # [Groups, OC * Tiles, IC / Tiles]
+            scales = scales[:, 0]  # [Groups * OC * Tiles, 1]
+            scales = scales.view(self.groups, -1,
+                                 n_tiles).transpose(1, 2)  # [Groups, Tiles, OC/Groups]
+        # else each tile has the same scaling factor (per-tensor or per-channel)
+        else:
+            scales = scales.view(self.groups, 1, -1)  # [Groups, 1, OC/Groups]
+            scales = scales.repeat(1, n_tiles, 1)  # [Groups, Tiles, OC/Groups]
+        thresholds *= scales  # translating centers back to the float range
+        return thresholds, scales
